+++ src/main/java/nextstep/qna/domain/BaseTime.java
@@ -1,0 +1,8 @@
+package nextstep.qna.domain;
+
+import java.time.LocalDateTime;
+
+public abstract class BaseTime {
+  protected final LocalDateTime createdDate = LocalDateTime.now();
+  protected LocalDateTime updatedDate;
+}
