+++ src/main/java/nextstep/qna/domain/DeleteHistory.java
@@ -1,11 +1,11 @@
 package nextstep.qna.domain;
 
+import nextstep.qna.domain.type.ContentType;
 import nextstep.users.domain.NsUser;
 
-import java.time.LocalDateTime;
 import java.util.Objects;
 
-public class DeleteHistory {
+public class DeleteHistory extends BaseTime {
     private Long id;
 
     private ContentType contentType;
@@ -14,16 +14,13 @@ public class DeleteHistory {
 
     private NsUser deletedBy;
 
-    private LocalDateTime createdDate = LocalDateTime.now();
-
     public DeleteHistory() {
     }
 
-    public DeleteHistory(ContentType contentType, Long contentId, NsUser deletedBy, LocalDateTime createdDate) {
+    public DeleteHistory(ContentType contentType, Long contentId, NsUser deletedBy) {
         this.contentType = contentType;
         this.contentId = contentId;
         this.deletedBy = deletedBy;
-        this.createdDate = createdDate;
     }
 
     @Override
@@ -45,6 +42,6 @@ public class DeleteHistory {
     @Override
     public String toString() {
         return "DeleteHistory [id=" + id + ", contentType=" + contentType + ", contentId=" + contentId + ", deletedBy="
-                + deletedBy + ", createdDate=" + createdDate + "]";
+                + deletedBy + ", createdDate=" + super.createdDate + "]";
     }
 }
