+++ src/main/java/nextstep/qna/domain/type/ContentType.java
@@ -1,4 +1,4 @@
-package nextstep.qna.domain;
+package nextstep.qna.domain.type;
 
 public enum ContentType {
     QUESTION, ANSWER;
